% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractODESolver.R
\docType{methods}
\name{AbstractODESolver,missing-method}
\alias{AbstractODESolver,missing-method}
\title{AbstractODESolver constructor `missing``}
\usage{
\S4method{AbstractODESolver}{missing}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
AbstractODESolver constructor `missing``
}
