% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractODESolver.R
\docType{methods}
\name{AbstractODESolver,ODE-method}
\alias{AbstractODESolver,ODE-method}
\title{AbstractODESolver constructor `ODE`}
\usage{
\S4method{AbstractODESolver}{ODE}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
Uses this constructor when ODE object is passed
}
