% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R
\docType{methods}
\name{setStepSize}
\alias{setStepSize}
\alias{setStepSize,AbstractODESolver-method,,ODESolver-method}
\alias{getStepSize}
\alias{step}
\alias{init,ODESolver-method}
\alias{setStepSize,ODESolver-method}
\alias{getStepSize,ODESolver-method}
\alias{step,ODESolver-method}
\alias{AbstractODESolver}
\alias{AbstractODESolver-class}
\alias{step,AbstractODESolver-method}
\alias{step,AbstractODESolver-methods}
\alias{setStepSize,AbstractODESolver-method}
\alias{getStepSize,AbstractODESolver-method}
\alias{AbstractODESolver}
\title{AbstractODESolver class}
\usage{
setStepSize(object, ...)

getStepSize(object, ...)

step(object, ...)

\S4method{init}{ODESolver}(object, stepSize, ...)

\S4method{setStepSize}{ODESolver}(object, stepSize, ...)

\S4method{getStepSize}{ODESolver}(object, ...)

\S4method{step}{ODESolver}(object, ...)

\S4method{step}{AbstractODESolver}(object, ...)

\S4method{setStepSize}{AbstractODESolver}(object, stepSize, ...)

\S4method{getStepSize}{AbstractODESolver}(object, ...)

AbstractODESolver(.ode)
}
\arguments{
\item{object}{the main class binder}

\item{...}{additional parameters}

\item{stepSize}{how much the solver should move}

\item{.ode}{the ODE object}
}
\description{
AbstractODESolver class

step(): advance the solver one step at a time

setStepSize():

AbstractODESolver constructor
}
\section{Slots}{

\describe{
\item{\code{stepSize}}{numeric.}

\item{\code{numEqn}}{numeric.}

\item{\code{ode}}{ODE.}
}}

