% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODESolverFactory.R
\docType{methods}
\name{createODESolver}
\alias{createODESolver}
\alias{createODESolver,ODESolverFactory-method}
\title{This is a factory method that creates an ODESolver using a name.}
\usage{
createODESolver(object, ...)

\S4method{createODESolver}{ODESolverFactory}(object, ode, solverName, ...)
}
\arguments{
\item{object}{an solver object}

\item{...}{an additional parameter}

\item{ode}{an ODE object}

\item{solverName}{the desired solver as a string}
}
\description{
This is a factory method that creates an ODESolver using a name.
}
