% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DTRK4}
\alias{DTRK4}
\title{Data table for RK4 results}
\format{a data frame derivative of activity}
\usage{
data(DTRK4)
}
\description{
Intervals are the columns, dates (rows), steps as values.
Data frame of 61x289
}
\keyword{datasets}
