% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{enableRuntimeExceptions}
\alias{enableRuntimeExceptions}
\title{enableRuntimeExceptions}
\usage{
enableRuntimeExceptions(object, enable, ...)
}
\arguments{
\item{object}{a class object}

\item{enable}{a boolean to enable exceptions}

\item{...}{additional parameters}
}
\description{
Enable Runtime Exceptions
}
\examples{
setMethod("enableRuntimeExceptions", "DormandPrince45", function(object, enable) {
    object@enableExceptions <- enable
})
}
