% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/Euler.R
\docType{class}
\name{Euler}
\alias{Euler}
\alias{Euler-class}
\alias{.Euler}
\alias{Euler,ODE-method}
\alias{Euler,missing-method}
\title{Euler class}
\usage{
Euler(ode, ...)

\S4method{Euler}{ODE}(ode, ...)

\S4method{Euler}{missing}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
Euler class

Euler constructor when `ODE` passed

Euler constructor `missing` is passed
}
