% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euler.R
\docType{methods}
\name{Euler,ODE-method}
\alias{Euler,ODE-method}
\title{# constructor ODE solver using Euler method}
\usage{
\S4method{Euler}{ODE}(ode, ...)
}
\description{
#' Euler constructor
#'
#' @export
euler <- function(.ode) {
    # Euler constructor
    .euler <- new("Euler", .ode)                     # create the Euler object
    .euler <- init(.euler, .euler@stepSize)            # iniialize Euler
    return(.euler)
}
}
