% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EulerRichardson.R
\docType{methods}
\name{EulerRichardson,ODE-method}
\alias{EulerRichardson,ODE-method}
\title{EulerRichardson constructor ODE}
\usage{
\S4method{EulerRichardson}{ODE}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
EulerRichardson constructor ODE
}
