% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{getEnergy}
\alias{getEnergy}
\title{Get the calculated energy level}
\usage{
getEnergy(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the calculated energy level
}
