% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{getErrorCode}
\alias{getErrorCode}
\title{getErrorCode}
\usage{
getErrorCode(object, tol, ...)
}
\arguments{
\item{object}{a class object}

\item{tol}{tolerance}

\item{...}{additional parameters}
}
\description{
Get an error code
}
\examples{
setMethod("getErrorCode", "DormandPrince45", function(object) {
return(object@error_code)
})
}
