% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{getODE}
\alias{getODE}
\title{getODE}
\usage{
getODE(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the ODE status from the solver
}
