% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{getRateCounts}
\alias{getRateCounts}
\title{Get the number of times that the rate has been calculated}
\usage{
getRateCounts(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the number of times that the rate has been calculated
}
