% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE.R
\docType{methods}
\name{getState}
\alias{getState}
\alias{getState,ODE-method}
\title{Gets the current state of the system}
\usage{
getState(object, ...)

\S4method{getState}{ODE}(object, ...)
}
\arguments{
\item{object}{the class object}

\item{...}{additional parameters}
}
\value{
a vector
}
\description{
A vector representing the state of the system
}
