% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/DormandPrince45.R
\docType{methods}
\name{getStepSize}
\alias{getStepSize}
\alias{getStepSize,ODESolver-method}
\alias{getStepSize,AbstractODESolver-method}
\alias{getStepSize,DormandPrince45-method}
\title{Get the step size}
\usage{
getStepSize(object, ...)

\S4method{getStepSize}{ODESolver}(object, ...)

\S4method{getStepSize}{AbstractODESolver}(object, ...)

\S4method{getStepSize}{DormandPrince45}(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the step size
}
