% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/Euler.R
\docType{methods}
\name{getStepSize}
\alias{getStepSize}
\alias{getStepSize,ODESolver-method}
\alias{getStepSize,AbstractODESolver-method}
\alias{getStepSize,Euler-method}
\title{Get the current size of the step}
\usage{
getStepSize(object, ...)

\S4method{getStepSize}{ODESolver}(object, ...)

\S4method{getStepSize}{AbstractODESolver}(object, ...)

\S4method{getStepSize}{Euler}(object, ...)
}
\arguments{
\item{object}{the main class binder}

\item{...}{additional parameters}
}
\description{
Get the current size of the step
}
