% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{getTime}
\alias{getTime}
\title{Get the elapsed time}
\usage{
getTime(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the elapsed time
}
