% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{getTolerance}
\alias{getTolerance}
\title{getTolerance}
\usage{
getTolerance(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the tolerance for the solver
}
