% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{importFromExamples}
\alias{importFromExamples}
\title{importFromExamples}
\usage{
importFromExamples(aClassFile, aFolder = "examples")
}
\arguments{
\item{aClassFile}{a file containing one or more classes}

\item{aFolder}{a folder where examples are located}
}
\description{
Source the R script
}
