% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODETest.R
\docType{class}
\name{ODETest-class}
\alias{ODETest-class}
\title{ODETest as an example of ODE class inheritance}
\description{
ODETest is a base class for examples ComparisonRK45App.R and
ComparisonRK45ODEApp.R
}
