% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE.R
\docType{class}
\name{ODE}
\alias{ODE}
\alias{ODE-class}
\alias{getRate}
\alias{getRate,ODE-method}
\alias{getRate,ODE-method}
\alias{getState}
\alias{getState,ODE-method}
\title{ODE class}
\usage{
getRate(object, state, rate, ...)

\S4method{getRate}{ODE}(object, state, rate, ...)

getState(object, ...)

\S4method{getState}{ODE}(object, ...)
}
\arguments{
\item{object}{the class object}

\item{state}{state of the ODE}

\item{rate}{rate of the ODE}

\item{...}{additional parameters}
}
\value{
.Object
}
\description{
ODE class

Gets the rate from the ODE solver after a step
}
\section{Slots}{

\describe{
\item{\code{state}}{numeric.}

\item{\code{rate}}{numeric.}
}}

