% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEAdaptiveSolver.R
\docType{class}
\name{ODEAdaptiveSolver-class}
\alias{ODEAdaptiveSolver-class}
\alias{.ODEAdaptiveSolver}
\alias{ODEAdaptiveSolver}
\alias{setTolerance,ODEAdaptiveSolver-method}
\alias{setTolerance<-,ODEAdaptiveSolver-method}
\alias{getTolerance,ODEAdaptiveSolver-method}
\alias{getErrorCode,ODEAdaptiveSolver-method}
\alias{ODEAdaptiveSolver,ANY-method}
\title{ODEAdaptiveSolver class}
\usage{
ODEAdaptiveSolver(...)

\S4method{setTolerance}{ODEAdaptiveSolver}(object, tol)

\S4method{setTolerance}{ODEAdaptiveSolver}(object, ...) <- value

\S4method{getTolerance}{ODEAdaptiveSolver}(object)

\S4method{getErrorCode}{ODEAdaptiveSolver}(object)

\S4method{ODEAdaptiveSolver}{ANY}(...)
}
\arguments{
\item{...}{additional parameters}

\item{object}{a class object}

\item{tol}{tolerance}

\item{value}{the value for the tolerance}
}
\description{
Base class to be inherited by adaptive solvers such as RK45

ODEAdaptiveSolver generic

ODEAdaptiveSolver constructor
}
