% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODEAdaptiveSolver.R
\docType{class}
\name{ODEAdaptiveSolver}
\alias{ODEAdaptiveSolver}
\alias{ODEAdaptiveSolver-class}
\alias{setTolerance}
\alias{setTolerance,ODEAdaptiveSolver-method}
\alias{getTolerance}
\alias{getTolerance,ODEAdaptiveSolver-method}
\alias{getErrorCode}
\alias{getErrorCode,ODEAdaptiveSolver-method}
\title{The ODEAdaptiveSolver class}
\usage{
setTolerance(object, tol, ...)

\S4method{setTolerance}{ODEAdaptiveSolver}(object, tol)

getTolerance(object, ...)

\S4method{getTolerance}{ODEAdaptiveSolver}(object)

getErrorCode(object, tol, ...)

\S4method{getErrorCode}{ODEAdaptiveSolver}(object)
}
\arguments{
\item{object}{a class object}

\item{tol}{tolerance}

\item{...}{additional parameters}
}
\description{
The ODEAdaptiveSolver class

Set the tolerance for the adaptive solver
}
\section{Slots}{

\describe{
\item{\code{NO_ERROR}}{numeric.}

\item{\code{DID_NOT_CONVERGE}}{numeric.}

\item{\code{BISECTION_EVENT_NOT_FOUND}}{numeric.}
}}

