% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODESolver.R, R/AbstractODESolver.R
\docType{class}
\name{ODESolver}
\alias{ODESolver}
\alias{ODESolver-class}
\alias{init}
\alias{init,ODESolver-method}
\alias{setStepSize}
\alias{setStepSize,ODESolver-method}
\alias{getStepSize}
\alias{getStepSize,ODESolver-class}
\alias{getStepSize,ODESolver-method}
\alias{getStepSize,ODESolver-class}
\alias{step}
\alias{step,ODESolver-method}
\alias{getStepSize,AbstractODESolver-method}
\alias{getStepSize,AbstractODESolver-class}
\title{ODESolver class}
\usage{
init(object, stepSize, ...)

\S4method{init}{ODESolver}(object, stepSize, ...)

setStepSize(object, ...)

\S4method{setStepSize}{ODESolver}(object, stepSize, ...)

getStepSize(object, ...)

\S4method{getStepSize}{ODESolver}(object, ...)

step(object, ...)

\S4method{step}{ODESolver}(object, ...)

\S4method{getStepSize}{AbstractODESolver}(object, ...)
}
\arguments{
\item{object}{the object}

\item{stepSize}{the size}

\item{...}{add}
}
\description{
ODESolver class

getStepSize()
}
