% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RK4.R
\docType{methods}
\name{RK4,ODE-method}
\alias{RK4,ODE-method}
\title{#' RK4 constructor
#'
#' @rdname RK4
#' @importFrom methods new
#' @export
rk4 <- function(.ode) {
    # constructor for RK4 ODE solver
    .rk4 <- new("RK4", .ode)
    .rk4 <- init(.rk4, .rk4@stepSize)
    return(.rk4)
}}
\usage{
\S4method{RK4}{ODE}(ode, ...)
}
\description{
#' RK4 constructor
#'
#' @rdname RK4
#' @importFrom methods new
#' @export
rk4 <- function(.ode) {
    # constructor for RK4 ODE solver
    .rk4 <- new("RK4", .ode)
    .rk4 <- init(.rk4, .rk4@stepSize)
    return(.rk4)
}
}
