% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{setSolver<-}
\alias{setSolver<-}
\title{setSolver}
\usage{
setSolver(object) <- value
}
\arguments{
\item{object}{a class object}

\item{value}{value to be set}
}
\description{
Set a solver over an ODE object
}
