% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/Euler.R
\docType{methods}
\name{setStepSize}
\alias{setStepSize}
\alias{setStepSize,AbstractODESolver-method,,ODESolver-method}
\alias{setStepSize,ODESolver-method}
\alias{setStepSize,AbstractODESolver-method}
\alias{setStepSize,Euler-method}
\title{Set the size of step to move each step of the solver}
\usage{
setStepSize(object, ...)

\S4method{setStepSize}{ODESolver}(object, stepSize, ...)

\S4method{setStepSize}{AbstractODESolver}(object, stepSize, ...)

\S4method{setStepSize}{Euler}(object, stepSize, ...)
}
\arguments{
\item{object}{the main class binder}

\item{...}{additional parameters}

\item{stepSize}{how much the solver should move}
}
\description{
Set the size of step to move each step of the solver
}
