% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/Euler.R
\docType{methods}
\name{step}
\alias{step}
\alias{step,ODESolver-method}
\alias{step,AbstractODESolver-method}
\alias{step,Euler-method}
\title{Advance the solver one step at a time}
\usage{
step(object, ...)

\S4method{step}{ODESolver}(object, ...)

\S4method{step}{AbstractODESolver}(object, ...)

\S4method{step}{Euler}(object, ...)
}
\arguments{
\item{object}{the main class binder}

\item{...}{additional parameters}
}
\description{
Advance the solver one step at a time
}
