% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/AbstractODESolver.R
\docType{class}
\name{AbstractODESolver}
\alias{AbstractODESolver}
\alias{AbstractODESolver-class}
\alias{.AbstractODESolver}
\title{AbstractODESolver class}
\usage{
AbstractODESolver(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
Defines the basic methods for all the ODE solvers.
}
\details{
Inherits from: ODESolver class
}
