% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractODESolver.R
\docType{methods}
\name{AbstractODESolver,missing-method}
\alias{AbstractODESolver,missing-method}
\title{# constructor
#' @importFrom methods new
AbstractODESolver <- function(.ode) {
    odesolver <- new("AbstractODESolver", .ode)
    odesolver@ode <- .ode
    odesolver
}}
\usage{
\S4method{AbstractODESolver}{missing}(ode, ...)
}
\description{
# constructor
#' @importFrom methods new
AbstractODESolver <- function(.ode) {
    odesolver <- new("AbstractODESolver", .ode)
    odesolver@ode <- .ode
    odesolver
}
}
