% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractODESolver.R
\docType{class}
\name{AbstractODESolver}
\alias{AbstractODESolver}
\alias{AbstractODESolver-class}
\alias{step,AbstractODESolver-method}
\alias{step,AbstractODESolver-methods}
\alias{setStepSize,AbstractODESolver-method}
\alias{init,AbstractODESolver-method}
\alias{AbstractODESolver}
\title{AbstractODESolver class}
\usage{
\S4method{step}{AbstractODESolver}(object, ...)

\S4method{setStepSize}{AbstractODESolver}(object, stepSize, ...)

\S4method{init}{AbstractODESolver}(object, stepSize, ...)

AbstractODESolver(.ode)
}
\arguments{
\item{object}{the main class binder}

\item{...}{additional parameters}

\item{stepSize}{how much the solver should move}

\item{.ode}{the ODE object}
}
\description{
AbstractODESolver class

step(): advance the solver one step at a time

setStepSize():

init(): not the same as initialize

AbstractODESolver constructor
}
\section{Slots}{

\describe{
\item{\code{stepSize}}{numeric.}

\item{\code{numEqn}}{numeric.}

\item{\code{ode}}{ODE.}
}}

