% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbstractODESolver.R
\docType{class}
\name{AbstractODESolver}
\alias{AbstractODESolver}
\alias{AbstractODESolver-class}
\alias{AbstractODESolver}
\title{AbstractODESolver class}
\usage{
AbstractODESolver(.ode)
}
\arguments{
\item{.ode}{an ODE object}
}
\description{
AbstractODESolver class

AbstractODESolver constructor
}
\section{Slots}{

\describe{
\item{\code{stepSize}}{numeric.}

\item{\code{numEqn}}{numeric.}

\item{\code{ode}}{ODE.}
}}

