% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/DormandPrince45.R
\docType{class}
\name{enableRuntimeExceptions}
\alias{enableRuntimeExceptions}
\alias{DormandPrince45-class}
\alias{enableRuntimeExceptions,DormandPrince45-method}
\title{Enable Runtime Exceptions}
\usage{
enableRuntimeExceptions(object, enable, ...)

\S4method{enableRuntimeExceptions}{DormandPrince45}(object, enable)
}
\arguments{
\item{object}{a class object}

\item{enable}{a logical flag}

\item{...}{additional parameters}
}
\description{
Enable Runtime Exceptions

DormandPrince45 class
}
