% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euler.R
\docType{class}
\name{Euler-class}
\alias{Euler-class}
\title{Euler class}
\description{
Euler implements an Euler method ODE solver.
}
\details{
The Euler method is unstable for many systems.  It is included as an
example of how to use the ODE and ODESolver interface.
}
\author{
Java: Wolfgang Christian

R: Alfonso R. Reyes
}
