% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Euler.R
\docType{methods}
\name{Euler,ODE-method}
\alias{Euler,ODE-method}
\title{# constructor ODE solver using Euler method}
\usage{
\S4method{Euler}{ODE}(ode, ...)
}
\description{
# constructor ODE solver using Euler method
}
