% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/EulerRichardson.R
\docType{class}
\name{EulerRichardson}
\alias{EulerRichardson}
\alias{EulerRichardson-class}
\alias{.EulerRichardson}
\title{EulerRichardson class}
\usage{
EulerRichardson(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
EulerRichardson class
}
