% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODEAdaptiveSolver.R,
%   R/DormandPrince45.R
\docType{methods}
\name{getErrorCode}
\alias{getErrorCode}
\alias{getErrorCode,ODEAdaptiveSolver-method}
\alias{getErrorCode,DormandPrince45-method}
\title{Get an error code}
\usage{
getErrorCode(object, tol, ...)

\S4method{getErrorCode}{ODEAdaptiveSolver}(object)

\S4method{getErrorCode}{DormandPrince45}(object)
}
\arguments{
\item{object}{a class object}

\item{tol}{tolerance}

\item{...}{additional parameters}
}
\description{
Get an error code
}
\examples{
setMethod("getErrorCode", "DormandPrince45", function(object) {
return(object@error_code)
})
}
\seealso{
Other adaptive solver methods: \code{\link{getTolerance}},
  \code{\link{setTolerance}}
}
