% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODETest.R
\docType{methods}
\name{getExactSolution}
\alias{getExactSolution}
\alias{getExactSolution,ODETest-method}
\title{Compare analytical and calculated solutions}
\usage{
getExactSolution(object, t, ...)

\S4method{getExactSolution}{ODETest}(object, t, ...)
}
\arguments{
\item{object}{a class object}

\item{t}{time ath what we are performing the evaluation}

\item{...}{additional parameters}
}
\description{
Compare analytical and calculated solutions
}
