% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R
\docType{methods}
\name{getODE}
\alias{getODE}
\alias{getODE,ODESolver-method}
\alias{getODE,AbstractODESolver-method}
\title{Get the ODE status from the solver}
\usage{
getODE(object, ...)

\S4method{getODE}{ODESolver}(object, ...)

\S4method{getODE}{AbstractODESolver}(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the ODE status from the solver
}
