% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODE.R, R/ODETest.R
\docType{methods}
\name{getRate}
\alias{getRate}
\alias{getRate,ODE-method}
\alias{getRate,ODETest-method}
\title{Get a new rate given a state}
\usage{
getRate(object, state, ...)

\S4method{getRate}{ODE}(object, state, ...)

\S4method{getRate}{ODETest}(object, state, ...)
}
\arguments{
\item{object}{a class object}

\item{state}{current state}

\item{...}{additional parameters}
}
\description{
Get a new rate given a state
}
