% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE.R
\docType{methods}
\name{getRate}
\alias{getRate}
\alias{getRate,ODE-method}
\title{Gets the current rate of the system}
\usage{
getRate(object, state, rate, ...)

\S4method{getRate}{ODE}(object, state, rate, ...)
}
\arguments{
\item{object}{the class object}

\item{state}{a vector holding the state of the ODE}

\item{rate}{a vector holding the rate}

\item{...}{additional parameters}
}
\description{
Given a state, return the rate calculated by the ODE solver after a step
}
