% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODETest.R
\docType{methods}
\name{getRateCounts}
\alias{getRateCounts}
\alias{getRateCounts,ODETest-method}
\title{Get the number of times that the rate has been calculated}
\usage{
getRateCounts(object, ...)

\S4method{getRateCounts}{ODETest}(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the number of times that the rate has been calculated
}
