% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODE.R, R/ODETest.R
\docType{methods}
\name{getState}
\alias{getState}
\alias{getState,ODE-method}
\alias{getState,ODETest-method}
\title{Get current state of the system}
\usage{
getState(object, ...)

\S4method{getState}{ODE}(object, ...)

\S4method{getState}{ODETest}(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get current state of the system
}
