% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODEAdaptiveSolver.R,
%   R/DormandPrince45.R
\docType{methods}
\name{getTolerance}
\alias{getTolerance}
\alias{getTolerance,ODEAdaptiveSolver-method}
\alias{getTolerance,DormandPrince45-method}
\title{Get the tolerance for the solver}
\usage{
getTolerance(object, ...)

\S4method{getTolerance}{ODEAdaptiveSolver}(object)

\S4method{getTolerance}{DormandPrince45}(object)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Get the tolerance for the solver
}
\seealso{
Other adaptive solver methods: \code{\link{getErrorCode}},
  \code{\link{setTolerance}}
}
