% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{init}
\alias{init}
\title{Set initial values before starting the ODE solver}
\usage{
init(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Set initial values before starting the ODE solver
}
