% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/Euler.R
\docType{methods}
\name{init}
\alias{init}
\alias{init-method}
\alias{init,ODESolver-method}
\alias{init,AbstractODESolver-method}
\alias{init,Euler-method}
\title{Sets the values of the ODE solver and gets ready to start}
\usage{
init(object, stepSize, ...)

\S4method{init}{ODESolver}(object, stepSize, ...)

\S4method{init}{AbstractODESolver}(object, stepSize, ...)

\S4method{init}{Euler}(object, stepSize, ...)
}
\arguments{
\item{object}{the main class binder}

\item{stepSize}{how much the solver should move}

\item{...}{additional parameters}
}
\description{
Sets the values of the ODE solver and gets ready to start

Start the Euler ODE solver
}
