% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODE.R
\docType{class}
\name{ODE-class}
\alias{ODE-class}
\alias{ODE}
\title{ODE class}
\description{
Creates an ODE object to be processed by an ODE solver
}
\section{Slots}{

\describe{
\item{\code{state}}{numeric.}

\item{\code{rate}}{numeric.}
}}

