% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ODESolver.R
\docType{class}
\name{ODESolver-class}
\alias{ODESolver-class}
\title{ODESolver virtual class}
\arguments{
\item{object}{a class object}

\item{stepSize}{size of the step}
}
\description{
A virtual class inherited by AbstractODESolver
}
