% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/RK4.R
\docType{class}
\name{RK4}
\alias{RK4}
\alias{RK4-class}
\alias{.RK4}
\alias{RK4,ODE-method}
\title{############################################################################}
\usage{
RK4(ode, ...)

\S4method{RK4}{ODE}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
RK4.R

RK4 class constructor
}
\details{
RK4 ODE solver

RK4 class
}
