% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{setSolver}
\alias{setSolver}
\title{Set a solver over an ODE object}
\usage{
setSolver(object, ode, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Set a solver over an ODE object
}
