% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODE.R
\docType{methods}
\name{setState}
\alias{setState}
\alias{setState}
\alias{setState,ODE-method}
\title{Set a new for the ODE solver}
\usage{
setState(object, ...)

setState(object, ...)

\S4method{setState}{ODE}(object, theta, thetaDot)
}
\arguments{
\item{object}{the class object}

\item{...}{additional parameters}

\item{theta}{angle}

\item{thetaDot}{derivative of angle}
}
\description{
New setState that should work with different methods

A vector representing the state of the system
}
\details{
"theta", "thetaDot":  used in PendulumApp
 "x", "vx", "y", "vy": used in ProjectileApp
}
