% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R
\name{setState}
\alias{setState}
\title{New setState that should work with different methods
 "theta", "thetaDot":  used in PendulumApp
 "x", "vx", "y", "vy": used in ProjectileApp}
\usage{
setState(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
New setState that should work with different methods
 "theta", "thetaDot":  used in PendulumApp
 "x", "vx", "y", "vy": used in ProjectileApp
}
