% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/DormandPrince45.R, R/Euler.R
\docType{methods}
\name{setStepSize}
\alias{setStepSize}
\alias{setStepSize,ODESolver-method}
\alias{setStepSize,AbstractODESolver-method}
\alias{setStepSize,DormandPrince45-method}
\alias{setStepSize,Euler-method}
\alias{getStepSize,Euler-method}
\title{setStepSize uses either of two step parameters: stepSize and dt
`stepSize`` works for most of the applications
`dt`` is used in Pendulum}
\usage{
setStepSize(object, ...)

\S4method{setStepSize}{ODESolver}(object, stepSize, ...)

\S4method{setStepSize}{AbstractODESolver}(object, stepSize, ...)

\S4method{setStepSize}{DormandPrince45}(object, stepSize, ...)

\S4method{setStepSize}{Euler}(object, stepSize, ...)

\S4method{getStepSize}{Euler}(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}

\item{stepSize}{size of the step}
}
\description{
setStepSize uses either of two step parameters: stepSize and dt
`stepSize`` works for most of the applications
`dt`` is used in Pendulum

Set the size of the step
}
