% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODEAdaptiveSolver.R,
%   R/DormandPrince45.R
\docType{methods}
\name{setTolerance}
\alias{setTolerance}
\alias{setTolerance,ODEAdaptiveSolver-method}
\alias{setTolerance,DormandPrince45-method}
\title{Set the tolerance for the solver}
\usage{
setTolerance(object, tol, ...)

\S4method{setTolerance}{ODEAdaptiveSolver}(object, tol)

\S4method{setTolerance}{DormandPrince45}(object, tol)
}
\arguments{
\item{object}{a class object}

\item{tol}{tolerance}

\item{...}{additional parameters}
}
\description{
Set the tolerance for the solver
}
