% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showMethods2}
\alias{showMethods2}
\title{showMethods2}
\usage{
showMethods2(theClass)
}
\arguments{
\item{theClass}{class to analyze}
}
\description{
Get the methods in a class. But only those specific to the class
}
