% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/ODESolver.R,
%   R/AbstractODESolver.R, R/DormandPrince45.R, R/Euler.R, R/EulerRichardson.R,
%   R/RK4.R, R/Verlet.R
\docType{methods}
\name{step}
\alias{step}
\alias{step,ODESolver-method}
\alias{step,AbstractODESolver-method}
\alias{step,DormandPrince45-method}
\alias{step,Euler-method}
\alias{step,EulerRichardson-method}
\alias{step,RK4-method}
\alias{step,Verlet-method}
\title{Advances a step in the ODE solver}
\usage{
step(object, ...)

\S4method{step}{ODESolver}(object, ...)

\S4method{step}{AbstractODESolver}(object, ...)

\S4method{step}{DormandPrince45}(object, ...)

\S4method{step}{Euler}(object, ...)

\S4method{step}{EulerRichardson}(object, ...)

\S4method{step}{RK4}(object, ...)

\S4method{step}{Verlet}(object, ...)
}
\arguments{
\item{object}{a class object}

\item{...}{additional parameters}
}
\description{
Advances a step in the ODE solver
}
