% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Verlet.R
\docType{class}
\name{Verlet-class}
\alias{Verlet-class}
\alias{.Verlet}
\title{############################################################################}
\description{
Verlet.R
}
\details{
Verlet ODE solver
}
