% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/Verlet.R
\docType{class}
\name{Verlet}
\alias{Verlet}
\alias{Verlet-class}
\alias{.Verlet}
\alias{Verlet,ODE-method}
\title{############################################################################}
\usage{
Verlet(ode, ...)

\S4method{Verlet}{ODE}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}
}
\description{
Verlet.R

Verlet class constructor ODE
}
\details{
Verlet ODE solver

Verlet class
}
