% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_generics.R, R/Verlet.R
\docType{class}
\name{Verlet}
\alias{Verlet}
\alias{getRateCounter}
\alias{Verlet-class}
\alias{.Verlet}
\alias{getRateCounter,Verlet-method}
\alias{Verlet,ODE-method}
\title{Get the rate counter}
\usage{
Verlet(ode, ...)

getRateCounter(object, ...)

\S4method{getRateCounter}{Verlet}(object, ...)

\S4method{Verlet}{ODE}(ode, ...)
}
\arguments{
\item{ode}{an ODE object}

\item{...}{additional parameters}

\item{object}{a class object}
}
\description{
How many time the rate has changed with step

Verlet.R

Verlet class constructor ODE
}
\details{
Verlet ODE solver

Verlet class
}
